LWDEV-6992  Switch to BigDecimal (#120)
- use big decimal
- fees use big decimal
- matching engine use big decimal
- use big decimal in services
- use big decimal in tests
- use bigdecimal in limitorderservice test
- use bigdecimal in rounding test
- use bigdecimal in stop limit test
- use bigdecimal in tests
- use bigdecimal in multilimitorder service test
- use bigdecimal in limit service test
- use bigdecimal market order and other tests
- all build errors are fixed
- fix limitorder service
- fix matching engine divide operation prevision with bigDecimal
- fixes
- fix market order tests
- implement big decimal serializer
- use setscale instead of parseDouble
- use big math lib for big decimal operations
- implement custom equals hashcode FeeTransfer
- implement equals hashcode
- remove not needed code
- fix import
- fix imports
- merge fixes
- fix perfromance tests
- use bean for stop order file db accessor
- AbstractFileOrderBookDatabaseAccessor use nio for handling dirs with many files
- deduplicate file name creation
- read old format of orders from order book
- fix nulls in asset pair
- refactoring
- review fixes
- create new fee instructions
- introduce new fee instructions
- add convertor for fee instructions
- fix logging
- fix merge

+++ src/main/kotlin/com/lykke/matching/engine/daos/Tick.kt
@@ -1,3 +1,5 @@
 package com.lykke.matching.engine.daos
 
-class Tick(var openPrice: Double, var closePrice: Double, var highPrice: Double, var lowPrice: Double)
+import java.math.BigDecimal
+
+class Tick(var openPrice: BigDecimal, var closePrice: BigDecimal, var highPrice: BigDecimal, var lowPrice: BigDecimal)
