+++ src/main/kotlin/com/lykke/matching/engine/services/ReservedBalanceUpdateService.kt
@@ -9,6 +9,7 @@ import com.lykke.matching.engine.outgoing.messages.BalanceUpdate
 import com.lykke.matching.engine.outgoing.messages.ClientBalanceUpdate
 import com.lykke.matching.engine.utils.NumberUtils
 import org.apache.log4j.Logger
+import java.math.BigDecimal
 import java.util.Date
 
 class ReservedBalanceUpdateService(private val balancesHolder: BalancesHolder) : AbstractService {
@@ -27,7 +28,7 @@ class ReservedBalanceUpdateService(private val balancesHolder: BalancesHolder) :
                 "reserved amount: ${NumberUtils.roundForPrint(message.reservedAmount)}")
 
         val balance = balancesHolder.getBalance(message.clientId, message.assetId)
-        if (message.reservedAmount > balance) {
+        if (BigDecimal.valueOf(message.reservedAmount) > balance) {
             messageWrapper.writeNewResponse(ProtocolMessages.NewResponse.newBuilder()
                     .setStatus(MessageStatus.BALANCE_LOWER_THAN_RESERVED.type))
             LOGGER.info("Balance (client ${message.clientId}, asset ${message.assetId}, ${NumberUtils.roundForPrint(balance)}) is lower that reserved balance ${NumberUtils.roundForPrint(message.reservedAmount)}")
@@ -36,7 +37,7 @@ class ReservedBalanceUpdateService(private val balancesHolder: BalancesHolder) :
 
         val currentReservedBalance = balancesHolder.getReservedBalance(message.clientId, message.assetId)
 
-        val updated = balancesHolder.updateReservedBalance(message.clientId, message.assetId, message.reservedAmount, false)
+        val updated = balancesHolder.updateReservedBalance(message.clientId, message.assetId, BigDecimal.valueOf(message.reservedAmount), false)
         if (!updated) {
             messageWrapper.writeNewResponse(ProtocolMessages.NewResponse.newBuilder().setStatus(MessageStatus.RUNTIME.type))
             LOGGER.info("Unable to save balance (client ${message.clientId}, asset ${message.assetId}, reserved ${NumberUtils.roundForPrint(message.reservedAmount)})")
@@ -45,7 +46,7 @@ class ReservedBalanceUpdateService(private val balancesHolder: BalancesHolder) :
         balancesHolder.sendBalanceUpdate(BalanceUpdate(message.uid,
                 MessageType.RESERVED_BALANCE_UPDATE.name,
                 Date(),
-                listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, balance, currentReservedBalance, message.reservedAmount)), messageWrapper.messageId!!))
+                listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, balance, currentReservedBalance,  BigDecimal.valueOf(message.reservedAmount))), messageWrapper.messageId!!))
 
         messageWrapper.writeNewResponse(ProtocolMessages.NewResponse.newBuilder()
                 .setStatus(MessageStatus.OK.type))
